Rename Editor.tsx to DocEditor.tsx to match the component name
